Update LaunchApp.java
The follow methods added updates the wakeupcall table, the valet table, and allows employees to view the bookings. However, the employee functionalilty, such as issuing the wake up calls, adding the cars into the valet, and view the booking need to be done through the employee interface, since this is functionality that should not be exposed through the guests.